Create arwen/common.go
