+++ arwen/interface.go
@@ -8,26 +8,6 @@ import (
 	vmcommon "github.com/ElrondNetwork/elrond-vm-common"
 )
 
-type StorageStatus int
-
-const (
-	StorageUnchanged StorageStatus = 0
-	StorageModified  StorageStatus = 1
-	StorageAdded     StorageStatus = 3
-	StorageDeleted   StorageStatus = 4
-)
-
-type BreakpointValue uint64
-
-const (
-	BreakpointNone        BreakpointValue = 0
-	BreakpointAbort       BreakpointValue = 1
-	BreakpointAsyncCall   BreakpointValue = 2
-	BreakpointSignalError BreakpointValue = 3
-	BreakpointSignalExit  BreakpointValue = 4
-	BreakpointOutOfGas    BreakpointValue = 5
-)
-
 type StateStack interface {
 	InitState()
 	PushState()
